+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Queries now don't share same caching callback builder instance. 
+
 ## [0.7.6] - 2021-08-17
 ### Fixed
 - PHP 8.1 compatibility issue ([#24](./issues/24)).